requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.2.0
Faker==37.4.0
favicon==0.7.0
flatbuffers==25.2.10
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
htbuilder==0.9.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kagglehub==0.3.12
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lxml==5.4.0
Markdown==3.8
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.43.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pooch==1.8.2
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
PyMatting==1.1.14
pymdown-extensions==10.15
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
rembg==2.0.66
requests==2.32.4
rpds-py==0.25.1
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
soupsieve==2.7
st-annotated-text==4.0.2
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.46.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
